moved files to appropriate space
