+++ react-whib/src/components/MapSearchControl.js
